// src/app/enums/project-type.enums.ts
export enum EProjectType {
  TM = 'T&M',
  FF = 'Fixed Frice',
  NB = 'Non-Bill',
  ODC = 'ODC',
  Product = 'Product',
  Training = 'Training',
  NoSalary = 'No Salary'
}
export enum EProjectTypeList {
  TM = 'T&M',
  FF = 'FF',
  NB = 'NB',
  ODC = 'ODC',
  Product = 'P',
  Training = 'T',
  NoSalary = 'NS'
}

export enum EProjectTypeId {
  TM = 0,
  FF = 1,
  NB = 2,
  ODC = 3,
  Product = 4,
  Training = 5,
  NoSalary = 6
}